Добавил закрытие формы. не смогу сделать чтобы при  нажатие на delet форма удалялас
